Update resource URLs and ignore some development files
Changed the URLs for external resources on index.html to point to a new directory. Also added .vs, .fleet, and .idea to .gitignore to prevent pushing IDE specific configurations and other unwanted changes to the code repository.

These changes are essential to keep the code clean and ensure the proper loading of resources from the correct paths.